+++ api/src/main/java/brooklyn/catalog/BrooklynCatalog.java
@@ -90,7 +90,9 @@ public interface BrooklynCatalog {
      * Fails if the same version exists in catalog.
      *
      * @throws IllegalArgumentException if the yaml was invalid
+     * @deprecated since 0.7.0 use {@link #addItems(String, boolean)}
      */
+    @Deprecated
     CatalogItem<?,?> addItem(String yaml);
     
     /**
@@ -100,7 +102,9 @@ public interface BrooklynCatalog {
      * item exists with the same symbolicName and version
      *
      * @throws IllegalArgumentException if the yaml was invalid
+     * @deprecated since 0.7.0 use {@link #addItems(String, boolean)}
      */
+    @Deprecated
     CatalogItem<?,?> addItem(String yaml, boolean forceUpdate);
     
     /**